Fixed randomly observed -ve bandwidth issue when bandwidth reaching  ~2GB.

+++ c0appz.c
@@ -123,7 +123,7 @@ static size_t write_data_to_file(FILE *fp, struct m0_bufvec *data,
 struct m0_realm uber_realm;
 unsigned unit_size = 0;
 int perf=0;				/* performance option 		*/
-extern int qos_total_weight; 		/* total bytes read or written 	*/
+extern int64_t qos_total_weight; 		/* total bytes read or written 	*/
 extern pthread_mutex_t qos_lock;	/* lock  qos_total_weight 	*/
 int trace_level=0;
 bool m0trace_on = false;
